requirements.txt
msoffcrypto-tool
comtypes
